requirements.txt
beautifulsoup4==4.13.3
google-auth==2.38.0
google-auth-oauthlib==1.0.0
gspread==6.2.0
pandas==2.1.4
psycopg2==2.9.10
requests==2.31.0
SQLAlchemy==2.0.39